+++ changelog.md
@@ -19,6 +19,7 @@ Codion Change Log
 ### is.codion.framework.domain
 - Column.Getter and Setter renamed GetValue and SetParameter respectively.
 - Entity.put() renamed set().
+- DerivedAttributeDefinition.Builder.ProviderStage added, related refactoring.
 ### is.codion.framework.model
 - EntityQueryModel.conditions() renamed condition().
 - EntityEditModel.EntityEditor.isNull(attribute) replaced with EditorValue.present().